ESC-1524 added new content for when scp08 has issues. Also cleaned up some flags and test
