+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/BusinessEntityEoriController.scala
@@ -66,8 +66,7 @@ class BusinessEntityEoriController @Inject() (
   }
 
   private val isEoriValid = Constraint[String] { eori: String =>
-    if (appConfig.xiEoriAddingDisabled && eori.replaceAll(" ", "").matches("""^(gb|Gb|gB|GB)[0-9]{12,15}$""")) Valid
-    else if (!appConfig.xiEoriAddingDisabled && eori.replaceAll(" ", "").matches(EORI.regex)) Valid
+    if (eori.replaceAll(" ", "").matches("""^(gb|Gb|gB|GB)[0-9]{12,15}$""")) Valid
     else Invalid("businessEntityEori.regex.error")
   }
