+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/FinancialDashboardController.scala
@@ -28,7 +28,7 @@ import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.FinancialDashboardPage
 import uk.gov.hmrc.eusubsidycompliancefrontend.views.models.FinancialDashboardSummary
 
 import javax.inject.{Inject, Singleton}
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 
 @Singleton
 class FinancialDashboardController @Inject() (
@@ -57,13 +57,10 @@ class FinancialDashboardController @Inject() (
         case None => throw new IllegalStateException(s"Undertaking for EORI: $eori not found")
       }
       subsidies <- escService.retrieveSubsidiesForDateRange(undertaking.reference, today.toSearchRange)
-      balanceOpt: Option[UndertakingBalance] <-
-        if (appConfig.scp08Enabled) escService.getUndertakingBalance(eori)
-        else Future.successful(None)
+      balanceOpt: Option[UndertakingBalance] <- escService.getUndertakingBalance(eori)
       summary = FinancialDashboardSummary.fromUndertakingSubsidies(undertaking, subsidies, balanceOpt, today)
     } yield
-      if (appConfig.scp08Enabled && balanceOpt.isEmpty) Redirect(routes.Scp08MaintenancePageController.showPage.url)
-      else if (undertaking.isManuallySuspended && appConfig.releaseCEnabled)
+      if (undertaking.isManuallySuspended)
         Redirect(routes.UndertakingSuspendedPageController.showPage(undertaking.isLeadEORI(eori)).url)
       else Ok(financialDashboardPage(summary))
