+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/SubsidyController.scala
@@ -685,7 +685,7 @@ class SubsidyController @Inject() (
           )
           _ <- store.update[SubsidyJourney](_.setSubmitted(true)).toContext
           _ <- escService.clearUndertakingCache(ref).toContext
-          isSuspended = appConfig.releaseCEnabled && undertaking.isAutoSuspended
+          isSuspended = undertaking.isAutoSuspended
         } yield Redirect(routes.SubsidyController.getClaimConfirmationPage(isSuspended))
 
         result.getOrElse(sys.error("Error processing subsidy cya form submission"))
