+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/FinancialDashboardPage.scala.html
@@ -18,7 +18,7 @@
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers.routes
 @import uk.gov.hmrc.eusubsidycompliancefrontend.views.formatters.BigDecimalFormatter.Syntax._
-@import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{H1, H2, P, scp08Banner}
+@import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{H1, H2, P}
 @import uk.gov.hmrc.eusubsidycompliancefrontend.views.models.FinancialDashboardSummary
 @import uk.gov.hmrc.eusubsidycompliancefrontend.util.TaxYearHelpers
 @import java.time.LocalDate
@@ -29,8 +29,7 @@
   govukTable: GovukTable,
   govukSummaryList: GovukSummaryList,
   govukInsetText: GovukInsetText,
-  govukNotificationBanner : GovukNotificationBanner,
-  scp08Banner: scp08Banner
+  govukNotificationBanner : GovukNotificationBanner
 )
 
 @(summary: FinancialDashboardSummary)(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
@@ -39,6 +38,12 @@
 
 @printDialog = {
     <div class="print-dialogue"><a class="govuk-link" href="#print-dialogue">@messages("common.print")</a></div>
+}
+
+@possibleMiscalculationContent = {
+    @if(summary.usingAllowanceRemaining){
+        @messages("dashboard.scp08Issues.text")
+    }
 }
 
 @mainSummary = @{
@@ -78,29 +83,16 @@
             )
         )
 
-    if(appConfig.scp08Enabled){
-        Seq(
-            SummaryListRow(
-                key = Key(
-                    content = HtmlContent(s"""<span id="undertaking-balance-heading">${messages(s"$key.undertaking.balance.heading")}</span>""")
-                ),
-                value = Value(
-                    content = HtmlContent(s"""<span id="undertaking-balance-value">${summary.undertakingBalanceEUR.toEuros}</span>""")
-                )
-            )
-        ) ++ mainPart
-    } else {
-        mainPart ++ Seq(
-            SummaryListRow(
-                key = Key(
-                    content = HtmlContent(s"""<span id="allowance-remaining-heading">${messages(s"$key.allowance.remaining.heading")}</span>""")
-                ),
-                value = Value(
-                    content = HtmlContent(s"""<span id="allowance-remaining-value">${summary.undertakingBalanceEUR.toEuros}</span>""")
-                )
+    Seq(
+        SummaryListRow(
+            key = Key(
+                content = HtmlContent(s"""<span id="undertaking-balance-heading">${messages(s"$key.undertaking.balance.heading")}</span>""")
+            ),
+            value = Value(
+                content = HtmlContent(s"""<span id="undertaking-balance-value">${summary.undertakingBalanceEUR.toEuros} ${possibleMiscalculationContent}</span>""")
             )
         )
-    }
+    ) ++ mainPart
 
 }
 
@@ -111,16 +103,13 @@
     printDialog = Some(printDialog)
 ) {
 
-    @scp08Banner()
     @H1(title)
     @P(messages(s"$key.p1"))
     @P(messages(s"$key.p2", TaxYearHelpers.taxYearStartForDate(LocalDate.now.minusYears(2)).getYear.toString))
 
-    @if(appConfig.scp08Enabled){
-        @govukInsetText(InsetText(
-            content = Text(messages("dashboard.insetText"))
-        ))
-    }
+    @govukInsetText(InsetText(
+        content = Text(messages("dashboard.insetText"))
+    ))
 
     @govukSummaryList(SummaryList(
         rows = mainSummary
