+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/LeadAccountPage.scala.html
@@ -18,7 +18,7 @@
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers.routes
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.Undertaking
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.EORI
-@import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{H1, H2, P, scp08Banner}
+@import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{H1, H2, P}
 
 @this(
   layout: Layout,
@@ -27,8 +27,7 @@
   govukWarningText : GovukWarningText,
   govukInsetText : GovukInsetText,
   govukDetails : GovukDetails,
-  govukNotificationBanner : GovukNotificationBanner,
-  scp08Banner: scp08Banner
+  govukNotificationBanner : GovukNotificationBanner
 )
 
 @(
@@ -46,7 +45,8 @@
     remainingAmount: String,
     currentPeriodStart: String,
     isOverAllowance: Boolean,
-    isSuspended: Boolean
+    isSuspended: Boolean,
+    usingAllowanceRemaining: Boolean
 )(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
 @key = @{"lead-account-homepage"}
 
@@ -56,8 +56,7 @@
 }
 
 @balanceHeading = @{
-    if(appConfig.scp08Enabled) messages("lead-account-homepage.h2.undertakingBalance")
-    else messages("lead-account-homepage.h2-4")
+    messages("lead-account-homepage.h2.undertakingBalance")
 }
 
 @detailsContent = {
@@ -67,13 +66,20 @@
     @P(messages("lead-account-homepage.details.neverSubmitted.h3.p1"), id = Some("lead-account-homepage-details-neverSubmitted-h3-p1"))
 }
 
+@possibleMiscalculationContent = {
+    @if(usingAllowanceRemaining){
+        @govukInsetText(InsetText(
+            content = Text(messages("account-homepage.scp08Issues.text")), id = Some("scp08Issues-text")
+        ))
+    }
+}
+
+
 @layout(
   pageTitle = Some(messages(s"$key.title")),
   backLinkEnabled = false,
   backLink = None
 ) {
-
-    @scp08Banner()
 
     @H1(messages("lead-account-homepage.title"), Some("govuk-heading-xl govuk-!-margin-bottom-0"))
     <p id="undertaking-eori-number" class="govuk-hint govuk-!-font-size-24 govuk-!-margin-top-1 govuk-!-margin-bottom-8">@eori</p>
@@ -131,6 +137,7 @@
     </ul>
 
     @H2(balanceHeading, Some("govuk-heading-m govuk-!-margin-top-4"), id = Some("undertaking-balance-section-heading"))
+    @possibleMiscalculationContent
     @P(messages(s"$key.p3", remainingAmount, allowance), id = Some("undertaking-balance-section-content"))
     <ul class="govuk-list govuk-list--spaced govuk-!-margin-bottom-8">
       <li><a class="govuk-link" href="@routes.FinancialDashboardController.getFinancialDashboard">@messages("lead-account-homepage.ul3-li1")</a></li>
