+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/NonLeadAccountPage.scala.html
@@ -18,15 +18,14 @@
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers.routes
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.Undertaking
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.EORI
-@import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{H1, H2, P, scp08Banner}
+@import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{H1, H2, P}
 
 @this(
   layout: Layout,
   button: components.Button,
   govukRadios : GovukRadios,
   govukWarningText : GovukWarningText,
-  govukInsetText : GovukInsetText,
-  scp08Banner: scp08Banner
+  govukInsetText : GovukInsetText
 )
 
 @(
@@ -40,13 +39,20 @@
     totalSubsidies: String,
     remainingAmount: String,
     currentPeriodStart: String,
-    isSuspended: Boolean
+    isSuspended: Boolean,
+    usingAllowanceRemaining: Boolean
 )(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
 @key = @{"non-lead-account-homepage"}
 
 @balanceHeading = @{
-    if(appConfig.scp08Enabled) messages(s"$key.h2.undertakingBalance")
-    else messages(s"$key.h2-2")
+    messages(s"$key.h2.undertakingBalance")
+}
+@possibleMiscalculationContent = {
+    @if(usingAllowanceRemaining){
+        @govukInsetText(InsetText(
+            content = Text(messages("account-homepage.scp08Issues.text")), id = Some("scp08Issues-text")
+        ))
+    }
 }
 
 @layout(
@@ -55,7 +61,6 @@
   backLink = None
 ) {
 
-  @scp08Banner()
   @H1(messages(s"$key.title"), Some("govuk-heading-xl govuk-!-margin-bottom-0"))
 
   <p class="govuk-hint govuk-!-font-size-24 govuk-!-margin-top-1 govuk-!-margin-bottom-8">@eori</p>
@@ -76,6 +81,7 @@
 
 
   @H2(balanceHeading, id = Some("undertaking-balance-section-heading"))
+  @possibleMiscalculationContent
   @P(messages(s"$key.p2", remainingAmount, allowance), id = Some("undertaking-balance-section-content"))
   
   <ul class="govuk-list govuk-list--spaced">
