--- app/uk/gov/hmrc/eusubsidycompliancefrontend/views/Scp08Page.scala.html
@@ -1,30 +1,0 @@
-@*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import uk.gov.hmrc.eusubsidycompliancefrontend.config.AppConfig
-@import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{P, H1}
-@import uk.gov.hmrc.eusubsidycompliancefrontend.controllers
-@import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.EmailStatus.Unverified
-
-@this(layout: Layout)
-@()(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
-
-@layout(pageTitle = Some(messages("scp08.maintenance-text.title"))) {
-  @H1(messages("scp08.maintenance-text.title"), Some("govuk-heading-xl govuk-!-margin-bottom-0"), id = Some("scp08-banner-title"))
-    &nbsp;
-  <p id="scp08-banner-p1" class="govuk-body">@Html(messages("scp08.maintenance-text.p1"))</p>
-  <p id="scp08-banner-p2" class="govuk-body">@Html(messages("scp08.maintenance-text.p2"))</p>
-}
