+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/helpers/P.scala.html
@@ -15,4 +15,4 @@
  *@
 
 @(content: String, classes: Option[String] = None, id: Option[String] = None)(implicit messages: play.api.i18n.Messages)
-<p class=@classes.getOrElse("govuk-body")@if(id){ id="@id" }>@Html(content)</p>
+<p @if(id.nonEmpty){ id="@id" } class='@classes.getOrElse("govuk-body")' >@Html(content)</p>
