+++ conf/application.conf
@@ -25,8 +25,6 @@ play.filters.enabled += play.filters.csp.CSPFilter
 # Default http client
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.HttpClientModule"
 
-play.modules.enabled += "StartModule"
-
 # Custom error handler
 play.http.errorHandler = "uk.gov.hmrc.eusubsidycompliancefrontend.config.ErrorHandler"
 
@@ -119,13 +117,6 @@ features{
   welsh-language-support = false
   # Change this value to true to enable the euro only journey
   euro-only-enabled = false
-  xi-eori-adding-disabled = false
-  scp08-enabled = true
-  release-c-enabled = false
-  run-remove-email-addresses-job-enabled = false
-  run-clear-exchange-rate-job-enabled = false
-  run-remove-removed-subsidies-job-enabled = false
-  run-remove-eori-email-store-job-enabled = true
 }
 
 # Replace play.i18n.langs with the commented out line below when your service has been fully translated into Welsh
