+++ test/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/AccountControllerSpec.scala
@@ -17,6 +17,8 @@
 package uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 
 import cats.implicits.catsSyntaxOptionId
+import org.jsoup.Jsoup
+import org.jsoup.nodes.Document
 import play.api.Configuration
 import play.api.inject.bind
 import play.api.inject.guice.GuiceableModule
@@ -25,12 +27,14 @@ import play.api.test.Helpers._
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.NilReturnJourney.Forms.NilReturnFormPage
 import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.{EligibilityJourney, NilReturnJourney, UndertakingJourney}
+import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.UndertakingStatus
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.{ConnectorError, NonHmrcSubsidy, Undertaking}
 import uk.gov.hmrc.eusubsidycompliancefrontend.persistence.Store
 import uk.gov.hmrc.eusubsidycompliancefrontend.services._
 import uk.gov.hmrc.eusubsidycompliancefrontend.syntax.FutureSyntax.FutureOps
 import uk.gov.hmrc.eusubsidycompliancefrontend.test.CommonTestData._
 import uk.gov.hmrc.eusubsidycompliancefrontend.util.TimeProvider
+import uk.gov.hmrc.eusubsidycompliancefrontend.syntax.TaxYearSyntax.LocalDateTaxYearOps
 
 import java.time.LocalDate
 import scala.concurrent.Future
@@ -57,8 +61,7 @@ class AccountControllerSpec
     Configuration.from(
       Map(
         // Disable CSP n=once hashes in rendered output
-        "play.filters.csp.nonce.enabled" -> false,
-        "features.scp08-enabled" -> false
+        "play.filters.csp.nonce.enabled" -> false
       )
     )
   )
@@ -83,7 +86,11 @@ class AccountControllerSpec
             mockRetrieveUndertaking(eori1)(undertaking.some.toFuture)
             mockGetOrCreate[EligibilityJourney](eori1)(Right(eligibilityJourneyComplete))
             mockGetOrCreate[UndertakingJourney](eori1)(Right(UndertakingJourney()))
-            mockRetrieveAllSubsidies(undertakingRef)(undertakingSubsidies.toFuture)
+            mockTimeProviderToday(fixedDate)
+            mockRetrieveSubsidiesForDateRange(undertakingRef, fixedDate.toSearchRange)(
+              undertakingSubsidies.toFuture
+            )
+            mockGetUndertakingBalance(eori1)(Future.successful(Some(undertakingBalance)))
             mockTimeProviderToday(fixedDate)
             mockGetOrCreate(eori1)(Right(nilJourneyCreate))
           }
@@ -104,10 +111,7 @@ class AccountControllerSpec
                 .getElementById("lead-account-homepage-p2-li2")
                 .text shouldBe "report that you have not been awarded any non-customs subsidy payments"
 
-              doc.getElementById("undertaking-balance-section-heading").text shouldBe "Remaining allowance"
-              doc
-                .getElementById("undertaking-balance-section-content")
-                .text shouldBe "Your undertaking currently has a remaining balance of €0.00, from your sector allowance of €12.34."
+              verifyUndertakingBalance(doc)
             }
           )
         }
@@ -131,9 +135,12 @@ class AccountControllerSpec
               undertaking.lastSubsidyUsageUpdt
                 .fold(List.empty[NonHmrcSubsidy])(d => List(nonHmrcSubsidy.copy(submissionDate = d)))
 
-            mockRetrieveAllSubsidies(undertakingRef)(
+            mockTimeProviderToday(currentDate)
+
+            mockRetrieveSubsidiesForDateRange(undertakingRef, currentDate.toSearchRange)(
               undertakingSubsidies.copy(nonHMRCSubsidyUsage = subsidies).toFuture
             )
+            mockGetUndertakingBalance(eori1)(Future.successful(Some(undertakingBalance)))
             mockTimeProviderToday(currentDate)
             mockGetOrCreate[NilReturnJourney](eori1)(Right(nilJourneyCreate))
           }
@@ -198,7 +205,11 @@ class AccountControllerSpec
               mockRetrieveUndertaking(eori4)(undertaking1.some.toFuture)
               mockGetOrCreate[EligibilityJourney](eori4)(Right(eligibilityJourneyComplete))
               mockGetOrCreate[UndertakingJourney](eori4)(Right(UndertakingJourney()))
-              mockRetrieveAllSubsidies(undertakingRef)(undertakingSubsidies.toFuture)
+              mockTimeProviderToday(fixedDate)
+              mockRetrieveSubsidiesForDateRange(undertakingRef, fixedDate.toSearchRange)(
+                undertakingSubsidies.toFuture
+              )
+              mockGetUndertakingBalance(eori4)(Future.successful(Some(undertakingBalance)))
               mockTimeProviderToday(fixedDate)
             }
 
@@ -211,14 +222,138 @@ class AccountControllerSpec
                 htmlBody should include regex routes.FinancialDashboardController.getFinancialDashboard.url
                 htmlBody should include regex routes.RemoveYourselfBusinessEntityController.getRemoveYourselfBusinessEntity.url
 
-                doc.getElementById("undertaking-balance-section-heading").text shouldBe "Remaining allowance"
-                doc
-                  .getElementById("undertaking-balance-section-content")
-                  .text shouldBe "Your undertaking currently has a remaining balance of €0.00, from your sector allowance of €12.34."
+                verifyUndertakingBalance(doc)
               }
             )
           }
 
+        }
+      }
+
+      "display the account home page with warning when undertaking is auto suspended" when {
+        "admin page loads home - 'undertakingStatus == active'" in {
+          val nilJourneyCreate = NilReturnJourney(NilReturnFormPage(None))
+          inSequence {
+            mockAuthWithEnrolmentAndNoEmailVerification()
+            mockRetrieveUndertaking(eori1)(
+              undertaking.copy(undertakingStatus = Some(UndertakingStatus.active)).some.toFuture
+            )
+            mockGetOrCreate[EligibilityJourney](eori1)(Right(eligibilityJourneyComplete))
+            mockGetOrCreate[UndertakingJourney](eori1)(Right(UndertakingJourney()))
+            mockTimeProviderToday(fixedDate)
+            mockRetrieveSubsidiesForDateRange(undertakingRef, fixedDate.toSearchRange)(
+              undertakingSubsidies.toFuture
+            )
+            mockGetUndertakingBalance(eori1)(Future.successful(Some(undertakingBalance)))
+            mockTimeProviderToday(fixedDate)
+            mockGetOrCreate(eori1)(Right(nilJourneyCreate))
+          }
+
+          val result = performAction()
+
+          val doc = Jsoup.parse(contentAsString(result))
+
+          verifyGenericHomepageContentForLead(doc)
+          verifyPreDeadlineContentForLead(doc)
+        }
+        "admin page loads home - 'undertakingStatus == suspendedAutomated'" in {
+          val nilJourneyCreate = NilReturnJourney(NilReturnFormPage(None))
+          inSequence {
+            mockAuthWithEnrolmentAndNoEmailVerification()
+            mockRetrieveUndertaking(eori1)(
+              undertaking.copy(undertakingStatus = Some(UndertakingStatus.suspendedAutomated)).some.toFuture
+            )
+            mockGetOrCreate[EligibilityJourney](eori1)(Right(eligibilityJourneyComplete))
+            mockGetOrCreate[UndertakingJourney](eori1)(Right(UndertakingJourney()))
+            mockTimeProviderToday(fixedDate)
+            mockRetrieveSubsidiesForDateRange(undertakingRef, fixedDate.toSearchRange)(
+              undertakingSubsidies.toFuture
+            )
+            mockGetUndertakingBalance(eori1)(Future.successful(Some(undertakingBalance)))
+            mockTimeProviderToday(fixedDate)
+            mockGetOrCreate(eori1)(Right(nilJourneyCreate))
+          }
+
+          val result = performAction()
+
+          val doc = Jsoup.parse(contentAsString(result))
+
+          verifyGenericHomepageContentForLead(doc)
+          verifyAutoSuspendContentForLead(doc)
+        }
+        "member page loads home" in {
+          inSequence {
+            mockAuthWithEnrolmentAndNoEmailVerification(eori4)
+            mockRetrieveUndertaking(eori4)(
+              undertaking1.copy(undertakingStatus = Some(UndertakingStatus.suspendedAutomated)).some.toFuture
+            )
+            mockGetOrCreate[EligibilityJourney](eori4)(Right(eligibilityJourneyComplete))
+            mockGetOrCreate[UndertakingJourney](eori4)(Right(UndertakingJourney()))
+            mockTimeProviderToday(fixedDate)
+            mockRetrieveSubsidiesForDateRange(undertakingRef, fixedDate.toSearchRange)(
+              undertakingSubsidies.toFuture
+            )
+            mockGetUndertakingBalance(eori4)(Future.successful(Some(undertakingBalance)))
+            mockTimeProviderToday(fixedDate)
+          }
+
+          val result = performAction()
+
+          val doc = Jsoup.parse(contentAsString(result))
+
+          doc.title() shouldBe "Your undertaking - Report and manage your allowance for Customs Duty waiver claims - GOV.UK"
+          doc
+            .getElementById("warning-text")
+            .text shouldBe "! Warning Your undertaking's deadline to submit a report passed on 20 April 2021."
+        }
+      }
+
+      "display the account home page with message about scp08 issues" when {
+        "admin page loads home" in {
+          val nilJourneyCreate = NilReturnJourney(NilReturnFormPage(None))
+          inSequence {
+            mockAuthWithEnrolmentAndNoEmailVerification()
+            mockRetrieveUndertaking(eori1)(
+              undertaking.some.toFuture
+            )
+            mockGetOrCreate[EligibilityJourney](eori1)(Right(eligibilityJourneyComplete))
+            mockGetOrCreate[UndertakingJourney](eori1)(Right(UndertakingJourney()))
+            mockTimeProviderToday(fixedDate)
+            mockRetrieveSubsidiesForDateRange(undertakingRef, fixedDate.toSearchRange)(
+              undertakingSubsidies.toFuture
+            )
+            mockGetUndertakingBalance(eori1)(Future.successful(None))
+            mockTimeProviderToday(fixedDate)
+            mockGetOrCreate(eori1)(Right(nilJourneyCreate))
+          }
+
+          val result = performAction()
+
+          val doc = Jsoup.parse(contentAsString(result))
+
+          verifyScp08IssuesMessage(doc)
+        }
+        "member page loads home" in {
+          inSequence {
+            mockAuthWithEnrolmentAndNoEmailVerification(eori4)
+            mockRetrieveUndertaking(eori4)(
+              undertaking1.some.toFuture
+            )
+            mockGetOrCreate[EligibilityJourney](eori4)(Right(eligibilityJourneyComplete))
+            mockGetOrCreate[UndertakingJourney](eori4)(Right(UndertakingJourney()))
+            mockTimeProviderToday(fixedDate)
+            mockRetrieveSubsidiesForDateRange(undertakingRef, fixedDate.toSearchRange)(
+              undertakingSubsidies.toFuture
+            )
+            mockGetUndertakingBalance(eori4)(Future.successful(None))
+            mockTimeProviderToday(fixedDate)
+          }
+
+          val result = performAction()
+
+          val doc = Jsoup.parse(contentAsString(result))
+
+          verifyScp08IssuesMessage(doc)
         }
       }
 
@@ -322,4 +457,63 @@ class AccountControllerSpec
     }
   }
 
+  private def verifyScp08IssuesMessage(doc: Document) = {
+    doc
+      .getElementById("scp08Issues-text")
+      .text shouldBe "Your undertaking balance may show a temporary miscalculation which will be fixed within 24 hours."
+  }
+
+  private def verifyUndertakingBalance(doc: Document): Unit = {
+    doc.getElementById("undertaking-balance-section-heading").text shouldBe "Undertaking balance"
+    doc
+      .getElementById("undertaking-balance-section-content")
+      .text shouldBe "Your undertaking currently has a remaining balance of €123.45, from your sector allowance of €12.34."
+  }
+
+  private def verifyAutoSuspendContentForLead(doc: Document) = {
+    doc
+      .getElementById("warning-text")
+      .text shouldBe "! Warning Your deadline to submit a report passed on 20 April 2021."
+    doc
+      .getElementById("lead-account-homepage-p1")
+      .text shouldBe "This date was 90 days after you either:"
+    doc
+      .getElementById("lead-account-homepage-p2")
+      .text shouldBe "You must now either:"
+    doc
+      .getElementById("lead-account-homepage-h2")
+      .text shouldBe "What you need to report now"
+    doc
+      .getElementById("lead-account-homepage-h2-p1")
+      .text shouldBe "You must use this service to report all non-customs subsidy payments you have received to continue claiming de minimis state aid."
+    doc
+      .getElementById("lead-account-homepage-h2-p2")
+      .text shouldBe "These payments can include grants and loans, provided as de minimis state aid from government and public authorities. You must report these at least once every 90 days, even if you have not received any payments."
+    doc
+      .getElementById("lead-account-homepage-h2-p3")
+      .text shouldBe "If you have received none, you must submit a report that you have received no payments for that period."
+  }
+
+  private def verifyPreDeadlineContentForLead(doc: Document) = {
+    doc
+      .getElementById("warning-text") shouldBe null
+    doc
+      .getElementById("lead-account-homepage-p1")
+      .text shouldBe "This date is 90 days after you either:"
+    doc
+      .getElementById("lead-account-homepage-p2")
+      .text shouldBe "You must either:"
+    doc
+      .getElementById("lead-account-homepage-h2")
+      .text shouldBe "What you need to report, and when"
+    doc
+      .getElementById("lead-account-homepage-h2-p1")
+      .text shouldBe "You must use this service to report all non-customs subsidy payments you have received."
+    doc
+      .getElementById("lead-account-homepage-h2-p2")
+      .text shouldBe "You must submit a report at least once every 90 days, even if you have not received any payments."
+    doc
+      .getElementById("lead-account-homepage-h2-p3")
+      .text shouldBe "If you do not submit a report, your undertaking may no longer be able to claim Customs Duty waivers."
+  }
 }
