+++ test/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/ControllerSpec.scala
@@ -122,16 +122,6 @@ trait ControllerSpec extends PlaySupport with ScalaFutures with IntegrationPatie
     }
   }
 
-  def verifyScp08Banner(document: Document) = {
-    document.getElementById("scp08-maintenance-banner") should not be null
-    document
-      .getElementById("scp08-banner-p1")
-      .text shouldBe "Please be aware we are currently experiencing some technical difficulties with the Customs Duty Waiver Scheme service, causing some traders to see a miscalculation of the undertaking balance. We are working urgently to resolve but if you are still within balance on the system and your records you can continue to submit supplementary declarations as normal and we will rectify the online balance. We apologise for any inconvenience caused."
-    document
-      .getElementById("scp08-banner-p2")
-      .text shouldBe "If you are experiencing issues preventing you from submitting your supplementary declaration then please contact - customs.duty-waivers@hmrc.gov.uk"
-  }
-
   def verifyGenericHomepageContentForLead(doc: Document, eori: EORI = CommonTestData.eori1) = {
     doc.title() shouldBe "Manage your undertaking - Report and manage your allowance for Customs Duty waiver claims - GOV.UK"
     doc.getElementById("undertaking-eori-number").text shouldBe eori
