+++ test/uk/gov/hmrc/eusubsidycompliancefrontend/controllers/LeadOnlyUndertakingSupportSpec.scala
@@ -31,7 +31,7 @@ import uk.gov.hmrc.eusubsidycompliancefrontend.models.Undertaking
 import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.EORI
 import uk.gov.hmrc.eusubsidycompliancefrontend.services.EscService
 import uk.gov.hmrc.eusubsidycompliancefrontend.syntax.FutureSyntax.FutureOps
-import uk.gov.hmrc.eusubsidycompliancefrontend.test.CommonTestData.{eori1, eori3, undertaking}
+import uk.gov.hmrc.eusubsidycompliancefrontend.test.CommonTestData.{eori1, eori3, manuallySuspendedUndertaking, undertaking}
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
 import uk.gov.hmrc.eusubsidycompliancefrontend.test.util.PlaySupport
@@ -59,7 +59,7 @@ class LeadOnlyUndertakingSupportSpec
 
       def runTest() = {
         val fakeRequest = authorisedRequestForEori(eori1)
-        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest, appConfig)
+        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest)
         status(result) shouldBe OK
       }
 
@@ -82,7 +82,7 @@ class LeadOnlyUndertakingSupportSpec
 
         val fakeRequest = authorisedRequestForEori(eori3)
 
-        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest, appConfig)
+        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest)
 
         status(result) shouldBe SEE_OTHER
         redirectLocation(result) should contain(routes.AccountController.getAccountPage.url)
@@ -95,10 +95,39 @@ class LeadOnlyUndertakingSupportSpec
 
         val fakeRequest = authorisedRequestForEori(eori1)
 
-        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest, appConfig)
+        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest)
 
         status(result) shouldBe SEE_OTHER
         redirectLocation(result) should contain(routes.AccountController.getAccountPage.url)
+      }
+    }
+
+    "redirect to the manually suspended page" when {
+
+      "called with a request from a manually suspended user who is a lead" in {
+        inSequence {
+          mockRetrieveUndertaking(eori1)(manuallySuspendedUndertaking.some.toFuture)
+        }
+        val fakeRequest = authorisedRequestForEori(eori1)
+
+        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest)
+
+        status(result) shouldBe SEE_OTHER
+        redirectLocation(result) should contain(routes.UndertakingSuspendedPageController.showPage(true).url)
+
+      }
+
+      "called with a request from a manually suspended user who is a non-lead" in {
+        inSequence {
+          mockRetrieveUndertaking(eori3)(manuallySuspendedUndertaking.some.toFuture)
+        }
+        val fakeRequest = authorisedRequestForEori(eori3)
+
+        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest)
+
+        status(result) shouldBe SEE_OTHER
+        redirectLocation(result) should contain(routes.UndertakingSuspendedPageController.showPage(false).url)
+
       }
     }
 
@@ -107,7 +136,7 @@ class LeadOnlyUndertakingSupportSpec
       def runTest() = {
         val fakeRequest = authorisedRequestForEori(eori1)
 
-        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest, appConfig)
+        val result = underTest.withLeadUndertaking(_ => Ok("Foo").toFuture)(fakeRequest)
 
         a[RuntimeException] shouldBe thrownBy(result.futureValue)
       }
